+++ modules/i18n/src/models/namespaces/flow-extension-ns.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com).
+ * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,21 +16,12 @@
  * under the License.
  */
 
-.flow-builder-execution {
-
-    &.apple,
-    &.github,
-    &.confirmation-code,
-    &.magic-link,
-    &.flow-extension {
-        img {
-            filter: invert(100%);
-        }
-    }
-
-    &.apple {
-        img {
-            color: white;
-        }
-    }
+export interface flowExtensionNS {
+    properties: {
+        description: string;
+        connectionLabel: string;
+        connectionPlaceholder: string;
+        noConnectionsWarning: string;
+        noConnectionsWarningWithSupport: string;
+    };
 }