+++ modules/i18n/src/models/namespaces/flows-ns.ts
@@ -120,6 +120,7 @@ export interface flowsNS {
                 facebook: string;
                 github: string;
                 google: string;
+                flowExtension: string;
                 microsoft: string;
                 passkeyEnrollment: string;
                 magicLink: string;
@@ -307,6 +308,10 @@ export interface flowsNS {
                     general: string;
                     noPoliciesAvailable: string;
                     policiesRequired: string;
+                };
+                flowExtension: {
+                    general: string;
+                    actionId: string;
                 };
             };
         };