+++ modules/i18n/src/translations/en-US/portals/index.ts
@@ -80,3 +80,4 @@ export * from "./common-users";
 export * from "./verifiable-credentials";
 export * from "./customer-data-service";
 export * from "./consents";
+export * from "./flow-extension";